Added check to enforce the name attribute of the checkBox tag
Added proper error handling to the checkBox tag in FormTagLib.
Previously, if the name attribute was not provided, the tag
assumed the name attribute was non-null and errored on an
indexOf call later in the tag. This commit adds a null check
to the name attribute and throws an appropriate error if
name is null