Grails Core: Refactor ArtefactHandlerAdapter and all the other ArtefactHandlers
Instead of using regular expressions to check the file path of a class,
the refactored solution is to directly check whether the file path matches the application directory.

At the same time, it clarifies the checking and judgment logic for Artefact at compile time and runtime.

* Using `isArtefact(ClassNode)` at compile time
* Using `isArtefact(Class)` at runtime

Update the all ArtefactHandlers,

* ArtefactHandlerAdapter
* ApplicationArtefactHandler
* BootstrapArtefactHandler
* CodecArtefactHandler
* ControllerArtefactHandler
* DomainClassArtefactHandler
* InterceptorArtefactHandler
* ServiceArtefactHandler
* TagLibArtefactHandler
* UrlMappingsArtefactHandler