Support using latest Grace Snapshot version to test
Snapshot is the latest, un-released version of Grace.It contains the most recent features and updates that are yet to be included in the official release.
Snapshot version is more suitable for developers who want to experiment with the latest features or contribute to the Grace codebase.

Now you can do it just use `create-app` command, there is no need to checkout from the source and build it first.

```
grace create-app grace.apps.blog --grace-version=2023.0.0-SNAPSHOT
```

Closes gh-424
See gh-423