Allow developers to customize the `app` directory
Grails employs a structured directory layout that organizes files and folders based on their functionality, which is essential for maintaining code in a large application.
The `app` directory is central to a Grails application, containing subdirectories for domains, services, views, controllers, taglibs and assets, which follow the MVC architectural pattern.
In the past, Grails has always adhered to the convention of using `grails-app` as a fixed directory since its inception. In Grace, we can use `grails-app`, or the more concise `app`.
Now, in the upcoming Grace 2024 version, we take it a step further by allowing developers to customize this directory. You are free to choose a more suitable specific directory, such as `web`, `admin`, `api`, or `store`.